fix(ofctrl): Uncertainty in Signal Transmission due to Multiple Parties Reading the Channel
In both the ofctrl and ofSwtich instances, they read from the
stream.Error channel, which has a capacity of only 1. When the
ovs-vswitchd connection breaks, the system will only write an error
signal to this channel once. As a result, only one of the two parties
can receive this error signal, and which one will be random. In the
previous design, when ofctrl received the error signal, it would only
send a signal to the disconnection channel shared with ofswitch. We have
improved this by having only ofswtitch receive the error signal, after
which ofswitch sends a disconnect signal to ofctrl. This resolves the
issue of potentially missing the signal in ofswitch.

Signed-off-by: echken <chengcheng.luo@smartx.com>